move programs around
