+++ packages/core/src/models/programs2/degree.test.ts
@@ -1,6 +1,6 @@
 import assert from 'assert'
 import test from 'node:test'
-import { withTestTransaction } from '../contexts/index.js'
+import { withTestTransaction } from '../../contexts/index.js'
 import { createDegree, listDegrees } from './degree.js'
 
 test('degree crud', () =>
