+++ packages/core/src/models/programs2/index.ts
@@ -1,0 +1,3 @@
+export * from './degree.js'
+export * from './discipline.js'
+export * from './program.js'
