+++ packages/core/src/models/programs2/program.ts
@@ -1,8 +1,7 @@
 import { schema } from '@nawadi/db'
 import { eq } from 'drizzle-orm'
-import { useQuery } from '../contexts/index.js'
-import { normalizeHandle, normalizeTitle } from './normalize.js'
-import { singleRow } from './row.js'
+import { useQuery } from '../../contexts/index.js'
+import { normalizeHandle, normalizeTitle, singleRow } from '../../util/index.js'
 
 export async function listPrograms() {
   const query = useQuery()
