+++ packages/core/src/util/normalize.ts
